feat(client): splitting out templates into their own folder and updating to match v1.0.0-alpha.1
