+++ packages/client/src/components/app/templates/index.ts
@@ -1,0 +1,6 @@
+export * from './templates.types';
+
+export { AskLLMTemplate } from './AskLLMTemplate';
+export { AskCSVTemplate } from './AskCSVTemplate';
+export { BlocksGuideTemplate } from './BlocksGuideTemplate';
+export { LandingPageTemplate } from './LandingPageTemplate';
