+++ packages/client/src/stores/state/state.types.ts
@@ -279,26 +279,3 @@ export type CellConfig<D extends CellDef = CellDef> = {
         parameters: D['parameters'],
     ) => string;
 };
-
-export type Template = {
-    /** Name of the template */
-    name: string;
-
-    /** Description associated with the template */
-    description: string;
-
-    /** State associated with the template */
-    state: SerializedState;
-
-    /** Image for the template */
-    image: string;
-
-    /** Author for the template */
-    author: string;
-
-    /** Date updated template */
-    lastUpdatedDate: string;
-
-    /** Tags associated with the template */
-    tags: string[];
-};
